+++ modules/apps/portal-template/portal-template-velocity/src/test/java/com/liferay/portal/template/velocity/internal/VelocityTemplateTest.java
@@ -27,7 +27,7 @@ import com.liferay.portal.kernel.util.HashMapBuilder;
 import com.liferay.portal.kernel.util.StringUtil;
 import com.liferay.portal.template.ClassLoaderResourceParser;
 import com.liferay.portal.template.TemplateContextHelper;
-import com.liferay.portal.template.velocity.configuration.VelocityEngineConfiguration;
+import com.liferay.portal.template.velocity.internal.configuration.VelocityEngineConfiguration;
 import com.liferay.portal.util.FileImpl;
 
 import java.io.IOException;
